USB lock up fix
This fixes the USB lock up when a host module is running a
GTK+ application on a slave.  This patch was provided by
Freescale

FSL SR 1-948563501

*changing the SBUSCFG register to 6. This instructs the USB
controller to use INCR8 bursts when possible to access the
memory and uses less bus bandwidth

*tune TXFIFOTHRES in the TXFILLTUNING register. This is
normally set to 2 which means the fifo is filled with 2
bursts (64 bytes) before the transfer on the USB bus is started.
Setting it to a higher value (Now 8 = 256 Bytes) will provide
more tolerance for latency on the memory bus and thus more less
risk for FIFO under run

Signed-off-by: Jeff Horn <jhorn@reachtech.com>