Ensure panel power/signal sequencing meets spec, eliminate white flash
This patch moves enabling LCD_DISP_ON out of the board file and into the
lcd_lms430 driver. Delays are added between enabling power to the panel,
the start of lcd signaling, the display enable, and the backlight enable
to prevent latchup or DC operation of the panel.

The chosen delays are midrange of each spec, or slightly over minimum where
only a minimum spec is given.

The backlight enable is (unfortunately) placed in the mxsfb driver rather
than the panel driver since we want to delay the backlight coming on
until the display subsystem is completely configured.

Signed-off-by: Steve Sakoman <steve@sakoman.com>