Add multiple panel support
This patch adds multiple panel support by extending
the ldb and fb device tree nodes so we can pass configuration
to the device drivers.

In order to support this there are now two dts files. Once we
have part numbers we can look at naming the dts files the same
as the hardware part number.

Added config option to select the LDB clock (PLL2, PLL5)
since some panels require a clock to low to be generated
from PLL2, the default.

Signed-off-by: Jeff Horn <jhorn@reachtech.com>