MLK-10782-1 ARM:imx6qdl:Add support to enable/disable MMDC_CH1 handshake
MMDC CH1 is not used on i.MX6Q, so the handshake needed to change the parent
of periph2_sel or the divider of mmdc_ch1_axi_podf will never succeed.
Disable the handshake mechanism to allow changing the frequency of mmdc_ch1_axi,
allowing to use it as a possible source for the LDB DI clock.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>
Signed-off-by: Ranjani Vaidyanathan <Ranjani.Vaidyanathan@freescale.com>
[commit 38236af1538294a766788e69936b920a4d394b1d from imx_3.14.38_6qp_ga]
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>