MLK-10782-2 ARM:imx: Set LDB_DI_CLK mux as read-only
Due to incorrect placement of the clock gate cell in the ldb_di[x]_clk tree,
the glitchy parent mux of ldb_di[x]_clk can cause a glitch to enter the
ldb_di_ipu_div divider. If the divider gets locked up, no ldb_di[x]_clk is
generated, and the LVDS display will hang when the ipu_di_clk is sourced from
ldb_di_clk.

To fix the problem, both the new and current parent of the ldb_di_clk should
be disabled before the switch. As this can not be guaranteed by the clock
framework during runtime, make the ldb_di[x]_sel muxes read-only. A workaround
to set the muxes once during boot could be added to the kernel or bootloader.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>
Signed-off-by: Ranjani Vaidyanathan <Ranjani.Vaidyanathan@freescale.com>
[commit 9e925f4c4deab86fe0115c1314404a6a833be9dd from imx_3.14.38_6qp_ga]
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>