ARM: dts: imx6qdl-g2h: Add pwm-beeper support
G2H solo board has a buzzer that is connected to two PWMs:

- PWM1 controls the frequency
- PWM2 controls the volume

Add support for 'pwm-beeper' so that standard Linux 'beep'
applicatios can work.

We currently use PWM1 as a GPIO output at 'high' level, which is configured
by the bootloader. This provides the buzzer with the maximum volume.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>